+++ app/Http/Controllers/api/v1/MeetingController.php
@@ -10,6 +10,7 @@ use App\Http\Resources\MeetingWithRoomAndServer as MeetingResource;
 use App\Models\Meeting;
 use App\Services\MeetingService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Log;
 
 /**
@@ -48,33 +49,37 @@ class MeetingController extends Controller
         if ($request->has('search') && trim($request->search) != '') {
             $searchQueries = explode(' ', preg_replace('/\s\s+/', ' ', $request->search));
             foreach ($searchQueries as $searchQuery) {
+                $searchQuery = strtolower($searchQuery);
                 $resource = $resource->where(function ($query) use ($searchQuery) {
                     $query->whereHas('room', function ($subQuery) use ($searchQuery) {
-                        $subQuery->where('name', 'like', '%'.$searchQuery.'%');
+                        $subQuery->where(DB::raw('LOWER(name)'), 'like', '%'.$searchQuery.'%');
                     })
                         ->orWhereHas('room.owner', function ($subQuery) use ($searchQuery) {
-                            $subQuery->where('firstname', 'like', '%'.$searchQuery.'%')
-                                ->orWhere('lastname', 'like', '%'.$searchQuery.'%');
+                            $subQuery->where(DB::raw('LOWER(firstname)'), 'like', '%'.$searchQuery.'%')
+                                ->orWhere(DB::raw('LOWER(lastname)'), 'like', '%'.$searchQuery.'%');
                         })
                         ->orWhereHas('server', function ($subQuery) use ($searchQuery) {
-                            $subQuery->where('name', 'like', '%'.$searchQuery.'%');
+                            $subQuery->where(DB::raw('LOWER(name)'), 'like', '%'.$searchQuery.'%');
                         });
                 });
             }
         }
 
-        // Sort table with allowed columns
-        if ($request->has('sort_by') && $request->has('sort_direction')) {
-            $by = $request->query('sort_by');
-            $dir = $request->query('sort_direction');
+        // Sort by column, fallback/default is start
+        $sortBy = match ($request->query('sort_by')) {
+            'room.participant_count' => 'room.participant_count',
+            'room.listener_count' => 'room.listener_count',
+            'room.voice_participant_count' => 'room.voice_participant_count',
+            'room.video_count' => 'room.video_count',
+            default => 'start',
+        };
 
-            if (in_array($by, ['start', 'room.participant_count', 'room.listener_count', 'room.voice_participant_count', 'room.video_count']) && in_array($dir, ['asc', 'desc'])) {
-                $resource = $resource->orderBy($by, $dir);
-            }
-        } else {
-            // Default ordering
-            $resource = $resource->orderBy('start');
-        }
+        // Sort direction, fallback/default is asc
+        $sortOrder = match ($request->query('sort_direction')) {
+            'desc' => 'DESC',
+            default => 'ASC',
+        };
+        $resource = $resource->orderByRaw($sortBy.' '.$sortOrder);
 
         // Respond with paginated result
         $resource = $resource->paginate(setting('pagination_page_size'));