+++ app/Http/Controllers/api/v1/RoomMemberController.php
@@ -14,6 +14,7 @@ use App\Models\Room;
 use App\Models\User;
 use Auth;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Log;
 
 class RoomMemberController extends Controller
@@ -34,14 +35,14 @@ class RoomMemberController extends Controller
 
         // Sort by column, fallback/default is firstname
         $sortBy = match ($request->query('sort_by')) {
-            'lastname' => 'lastname',
-            default => 'firstname',
+            'lastname' => 'LOWER(lastname)',
+            default => 'LOWER(firstname)',
         };
 
         // Sort direction, fallback/default is asc
         $sortOrder = match ($request->query('sort_direction')) {
-            'desc' => 'desc',
-            default => 'asc',
+            'desc' => 'DESC',
+            default => 'ASC',
         };
 
         // Filter by role, fallback/default is no filter
@@ -53,7 +54,7 @@ class RoomMemberController extends Controller
         };
 
         // Get all members of the room and sort them
-        $resource = $room->members()->orderBy($sortBy, $sortOrder);
+        $resource = $room->members()->orderByRaw($sortBy.' '.$sortOrder);
 
         // count all before applying filters
         $additional['meta']['total_no_filter'] = $resource->count();
@@ -63,9 +64,10 @@ class RoomMemberController extends Controller
             // Split search query into single words and search for them in firstname and lastname
             $searchQueries = explode(' ', preg_replace('/\s\s+/', ' ', $request->search));
             foreach ($searchQueries as $searchQuery) {
+                $searchQuery = strtolower($searchQuery);
                 $resource = $resource->where(function ($query) use ($searchQuery) {
-                    $query->where('firstname', 'like', '%'.$searchQuery.'%')
-                        ->orWhere('lastname', 'like', '%'.$searchQuery.'%');
+                    $query->where(DB::raw('LOWER(firstname)'), 'like', '%'.$searchQuery.'%')
+                        ->orWhere(DB::raw('LOWER(lastname)'), 'like', '%'.$searchQuery.'%');
                 });
             }
         }
@@ -102,7 +104,7 @@ class RoomMemberController extends Controller
     public function bulkImport(Room $room, BulkImportRequest $request)
     {
         foreach ($request->user_emails as $userEmail) {
-            $user = User::firstWhere('email', $userEmail);
+            $user = User::firstWhere(DB::raw('LOWER(email)'), strtolower($userEmail));
             $room->members()->attach($user, ['role' => $request->role]);
         }