+++ tests/Fixtures/validDatabase-postgres.sql
@@ -1,0 +1,6 @@
+CREATE TABLE test (
+                      "column-a" integer NOT NULL,
+                      "column-b" integer NOT NULL,
+                      "column-c" integer NOT NULL,
+                      "column-d" integer NOT NULL
+);